Support for specifying ceph daemons to be deployed
Currently, deploying ceph using kolla-ceph will include all
enabled daemons, such as mon, mgr, osd, rgw and mds, etc. But the
disadvantage is that modifying one daemon may affect all other
daemons.
For example, when osd is broken, the user only wants to redeploy osd
and does not want to redeploy other daemons.

This function can be used with --limit (limit the node that the task
executes), so that you can modify a daemon on a node.