Support ceph device classes feature
Device classes[1] is a new property of OSD. This feature allows the
data in the ceph pool to be placed in the specified OSDs. Only need
to specify the corresponding device class in the crush rule.

This commit supports the device classes feature and updates the steps
to create a pool.

[1]https://ceph.com/community/new-luminous-crush-device-classes/