Support serial mode to upgrade ceph
When the ceph image tag changes, such as adding new features or ceph
version upgrades, users need to upgrade the old version of the image
to the new version.
This commit provides a safer way to upgrade, serial mode, it has the
following features:

1. Only upgrade one ceph daemon on one node at a time.
2. Perform ceph status detection before upgrading each component,
upgrade only in HEALTH_OK state.
3. When upgrading osd, noout and norebalance are set before the upgrade
to prevent data rebalancing. The osd upgrade on the same node is also
run in sequence.