Support multipath disk for bluestore
1. Support for bluestore osd deployment on multipath disks.
2. Because the symbolic link of the multipath disk uses the format
of 'typecode.partuuid', in order to maintain consistency, the symbolic
link of the regular disk changes from partlabel to partuuid.
3. There are many types of disk naming suffixes that are 'p+number',
such as the default naming of multipath disks, bcache disks, but the
script currently only supports loop devices, this patch fixes this
problem.
4. Optimized the return information of find_disks.py, only returns
valid disk information, does not return null values.
5. Kolla-ceph supports calculating the weight of osd according to the size
of the disk. The current script uses the osd data partition(100M) when
calculating the weight of the bluestore osd. The block partition
should be used. This patch fixes this problem.